Warn if hFILE function return statuses are ignored
Ignoring the return codes from hread()/hwrite() etc leads to reliability
problems as failures in writing data will not be reported to the user.
Especially because this is internal code, use attributes to cause
compilation warnings if the error returns from these functions are
not checked.

Eventually we may have a thin hts_types.h or hts_utils.h header or
similar, to which the definition of HTS_RESULT_USED should migrate.

For now, ignoring errors from hgetc(), hputc(), and hputs() is permitted
as one may like to batch up the error checking up for these, but this
may be revisited later.