Maintain our own hFILE offset within the stream
Code using htell() expects to be able to call it a lot, so causing an
lseek(2) system call each time is not ideal.  Further, lseek(2) gives the
file position of the end/start of the read/write buffer (respectively),
so needs correcting to get the proper logical file position anyway.

With this change, hFILE_backend::tell() is no longer needed.