Add libcurl hFILE backend
This code is a reworking of Heng Li's kurl.c (PR #164) into an hFILE
backend.  Improvements include leaving buffering to the existing
buffering in hfile.c, full error checking, and some bug fixes.

SEEK_CUR is not supported as the current file position is not tracked:
possibly it could be derived from CURLINFO_SIZE_DOWNLOAD and the most
recently-set CURLOPT_RESUME_FROM position, but this is probably more
trouble than it's worth.  (Currently hseek(SEEK_CUR) results in a
backend SEEK_SET, so this is not a limitation.)

The single shared curl.multi might be more trouble than it's worth.
If checking nrunning proves to be unreliable or having a single CURLM
causes multi-threading trouble, we may need to have a CURLM in each
hFILE_libcurl instead.  (If there were a form of CURL_WRITEFUNC_PAUSE
that caused curl_easy_perform() to return early, we wouldn't need to
use the multi interface at all!  However this is unlikely as it's a
significant change to how curl_easy_perform() works.)

Soon to follow: writing; handling of s3://bucket/filepath URLs.