Add versioning for this plugin collection
As these are plugins, moreover separately-maintained plugins, it is
incorrect to use hts_version() to report HTSlib's version number,
as a compiled plugin could be used with a variety of HTSlib versions.

Compute a version number from the Git description of the working tree
(or a fallback if we are not a Git repository); as most plugins have
just one substantial source file, we can do this in a simpler way
than the version.h machinery used in the htslib repository.

Use string literal pasting rather than printf() as this puts the string
in the object file in one piece, so it can be observed with strings(1).