Simplify IRODS_VERSION_INTEGER handling; add another -L option
If IRODS_VERSION_INTEGER is undefined, detect whether we're using 3.x
or 4.0.x and define it accordingly.  Fail immediately with #error if
the iRODS SDK is a known-broken version.

If it exists look for libraries in $(IRODS_HOME)/lib/development_libraries,
which is populated by packaging/make_irods_dev_archive.sh in 4.1.x at least.