Alter non-Git version number to follow tag convention
We now occasionally tag vaguely-stable commits with YYYYMM tags.
This repository is a collection of unrelated plugins, so this is not
a "release" as such, but having a tag provides an immutable ref that
is useful for various purposes.  Hat tip @keithj.  It is tempting to
also remove --always from the git describe command, but I suppose
some people may not fetch tags.

For untagged commits, the User-Agent will be htslib-irods/YYYYMM-N-gHASH
and similarly for other plugins, so to avoid confusion we don't use
hyphens in the tag name or non-Git version number.